+++ src/Sut.WinForms.ScreenComponentsTest/ObjectRepository/RebasedUpperRightComponent.cs
@@ -7,13 +7,13 @@ namespace Sut.WinForms.ScreenComponentsTest.ObjectRepository
     public class RebasedUpperRightComponent : ScreenComponent<WinGroup>
     {
         public RebasedUpperRightComponent()
-            : base(By.Name("Upper Right Group"))
+            : base(By.ControlName("groupBoxUpperRight"))
         {
         }
 
         public bool CheckBoxExists
         {
-            get { return Find<WinCheckBox>(By.Name("Upper right control")).Exists; }
+            get { return Find<WinCheckBox>(By.ControlName("checkBoxUpperRight")).Exists; }
         }
     }
 }